Changes to about.html, choice.html, and dumb.css
Modified the about page to make it both look and read less awful.

Started work on the choice.html file, added info. Turns out MCIS, the place I'm supposed to get data from, doesn't work on Firefox for Linux. How lovely of them.

Modified the CSS file to make the <dt> tag bold. This adds to the seperation effect that is seen in the <dl> lists