Freeze on CSS
The CSS for this project has been finalized. Unless a small change is needed, the CSS for the project is complete for the time being. This is gonna be abandoned after it is due, this isn't a project I wanted to work on. I'd rather be working on RomeBot or something right now.

On to the changes, I removed the commenting that has been brought over from copy and pastes for a more consistent look. I'm sure I will regret that later.